se extrae en componentes el codigo
