+++ src/components/carlosfernandezcabrero/Header.astro
@@ -1,4 +1,4 @@
-<header class='py-3 px-4 mb-14 flex items-center justify-center sm:justify-end'>
+<header class='py-3 px-4 mb-12 flex items-center justify-center sm:justify-end'>
 	<a
 		href='https://github.com/carlosfernandezcabrero/password-generator.git'
 		target='_blank'
